pyproject.toml
[project]
name = "pyladies-global-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=6.0.0",
    "faicons>=0.2.2",
    "ipyleaflet>=0.19.2",
    "maplibre[all]>=0.3.5",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "plotly>=6.1.1",
    "plotnine>=0.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "shiny>=1.5.0",
    "shinylive>=0.8.4",
    "shinywidgets>=0.6.2",
    "tqdm>=4.67.1",
]
